+++ standalone/UgripperRuntime/include/record_runtime.h
@@ -67,7 +67,7 @@ public:
 
     bool initialize();
     int run();
-    void requestStop();
+    void requestStop() noexcept;
 
 private:
     struct ButtonSnapshot